backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
azure-cognitiveservices-speech==1.47.0
openai==1.54.0
python-dotenv==1.0.0
reportlab==4.2.5
Pillow==11.0.0
pydub==0.25.1
gunicorn==21.2.0
